x/capy/types/messages_animals_test.go
package types

import (
	"testing"

	"capy/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgCreateAnimals_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateAnimals
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateAnimals{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateAnimals{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgUpdateAnimals_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgUpdateAnimals
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgUpdateAnimals{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgUpdateAnimals{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgDeleteAnimals_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgDeleteAnimals
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgDeleteAnimals{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgDeleteAnimals{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
